Add test coverage reporting and add tests
