+++ .coveragerc
@@ -1,0 +1,3 @@
+[run]
+omit =
+    */tests/*
