+++ .gitignore
@@ -1,3 +1,4 @@
 data/
 darts/docs/
 .DS_Store
+.coverage
