+++ requirements.txt
@@ -4,3 +4,4 @@ pdftotext==3.0.0
 dependency-injector==4.44.0
 selenium
 python-dateutil
+coverage
